Move pre 1.5.0 unit tests into the unit test target
